de-duped code, renamed files
