+++ scripts/ci/determine-tests.mjs
@@ -3,7 +3,7 @@ import {setOutput} from '@actions/core';
 import {readdirSync, statSync} from 'fs';
 import {EOL} from 'os';
 import {basename, dirname, join, relative} from 'path';
-import {getBaseHeadSHAs, getChangedFiles} from './hasFileChanged.mjs';
+import {getBaseHeadSHAs, getChangedFiles} from './git-utils.mjs';
 import {listImports, ensureFileExists} from './list-imports.mjs';
 
 class NoRelevantChangesError extends Error {
