+++ scripts/ci/git-utils.mjs
@@ -1,0 +1,24 @@
+import {context} from '@actions/github';
+import {execSync} from 'node:child_process';
+
+export function getBaseHeadSHAs() {
+  switch (context.eventName) {
+    case 'pull_request':
+      return {
+        base: context.payload.pull_request.base.sha,
+        head: context.payload.pull_request.head.sha,
+      };
+    case 'merge_group':
+      return {
+        base: context.payload.merge_group.base_sha,
+        head: context.payload.merge_group.head_sha,
+      };
+  }
+}
+
+export function getChangedFiles(from, to) {
+  return execSync(`git diff --name-only ${from}..${to}`, {
+    stdio: 'pipe',
+    encoding: 'utf-8',
+  });
+}
